requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
colorama==0.4.4
cryptography==36.0.0
cycler==0.10.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.46.0
django-crispy-forms==1.13.0
django-on-heroku==1.1.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
whitenoise==5.3.0
